+++ src/com/landawn/abacus/util/Multimap.java
@@ -1023,16 +1023,6 @@ public class Multimap<K, E, V extends Collection<E>> {
         }
     }
 
-    public <X extends Exception> void flatForEach(Try.Consumer<? super E, X> action) throws X {
-        N.checkArgNotNull(action);
-
-        for (V v : valueMap.values()) {
-            for (E e : v) {
-                action.accept(e);
-            }
-        }
-    }
-
     public <X extends Exception> void flatForEach(Try.BiConsumer<? super K, ? super E, X> action) throws X {
         N.checkArgNotNull(action);
 
@@ -1043,6 +1033,43 @@ public class Multimap<K, E, V extends Collection<E>> {
 
             for (E e : entry.getValue()) {
                 action.accept(key, e);
+            }
+        }
+    }
+
+    public <X extends Exception> void forEachKey(final Try.Consumer<? super K, X> action) throws X {
+        N.checkArgNotNull(action);
+
+        for (K k : valueMap.keySet()) {
+            action.accept(k);
+        }
+    }
+
+    public <X extends Exception> void forEachValue(final Try.Consumer<? super V, X> action) throws X {
+        N.checkArgNotNull(action);
+
+        for (V v : valueMap.values()) {
+            action.accept(v);
+        }
+    }
+
+    /**
+     * 
+     * @param action
+     * @throws X
+     * @deprecated replaced by {@code flatForEachValue(com.landawn.abacus.util.Try.Consumer)}
+     */
+    @Deprecated
+    public <X extends Exception> void flatForEach(Try.Consumer<? super E, X> action) throws X {
+        flatForEachValue(action);
+    }
+
+    public <X extends Exception> void flatForEachValue(Try.Consumer<? super E, X> action) throws X {
+        N.checkArgNotNull(action);
+
+        for (V v : valueMap.values()) {
+            for (E e : v) {
+                action.accept(e);
             }
         }
     }